feat(engine): extract mdd seeding helpers
