+++ packages/engine/package.json
@@ -6,6 +6,8 @@
   "main": "./src/index.js",
   "types": "./src/index.d.ts",
   "dependencies": {
-    "tslib": "^2.3.0"
+    "tslib": "^2.3.0",
+    "date-fns": "^4.1.0",
+    "csv-parse": "^5.6.0"
   }
 }
