+++ packages/engine/src/index.ts
@@ -1,1 +1,2 @@
 export * from './lib/engine';
+export * from './lib/mdd-loader';
